+++ cmd/internal/docker.go
@@ -1,10 +1,14 @@
 package internal
 
 import (
+	"archive/tar"
+	"compress/gzip"
 	"context"
 	"encoding/binary"
 	"fmt"
+	"io"
 	"log"
+	"os"
 	"path/filepath"
 	"strings"
 	"time"
@@ -30,7 +34,7 @@ var (
 	dockerCmd = "docker"
 )
 
-// Container is a custom type for storing container information similar to output from “docker containers ls“.
+// Container is a custom type for storing container information similar to output from "docker containers ls".
 type Container struct {
 	ID     string
 	Image  string
@@ -57,7 +61,7 @@ func (c Containers) Swap(i, j int) {
 	c[i], c[j] = c[j], c[i]
 }
 
-// EvaluateDockerComposeStatus determines if the host has the “docker compose“ plugin or the “docker compose“
+// EvaluateDockerComposeStatus determines if the host has the "docker compose" plugin or the "docker compose"
 // script installed and set the global `dockerCmd` variable.
 func EvaluateDockerComposeStatus() error {
 	fmt.Println("[+] Checking the status of Docker and the Compose plugin...")
@@ -96,8 +100,8 @@ func EvaluateDockerComposeStatus() error {
 	return nil
 }
 
-// RunDockerComposeInstall executes the “docker compose“ commands for a first-time installation with
-// the specified YAML file (“yaml“ parameter).
+// RunDockerComposeInstall executes the "docker compose" commands for a first-time installation with
+// the specified YAML file ("yaml" parameter).
 func RunDockerComposeInstall(yaml string) {
 	buildErr := RunCmd(dockerCmd, []string{"-f", yaml, "build"})
 	if buildErr != nil {
@@ -138,8 +142,8 @@ func RunDockerComposeInstall(yaml string) {
 	fmt.Println("[+] You can get your admin password by running: ghostwriter-cli config get admin_password")
 }
 
-// RunDockerComposeUpgrade executes the “docker compose“ commands for re-building or upgrading an
-// installation with the specified YAML file (“yaml“ parameter).
+// RunDockerComposeUpgrade executes the "docker compose" commands for re-building or upgrading an
+// installation with the specified YAML file ("yaml" parameter).
 func RunDockerComposeUpgrade(yaml string, skipseed bool) {
 	fmt.Printf("[+] Running `%s` commands to build containers with %s...\n", dockerCmd, yaml)
 	downErr := RunCmd(dockerCmd, []string{"-f", yaml, "down"})
@@ -172,8 +176,8 @@ func RunDockerComposeUpgrade(yaml string, skipseed bool) {
 	fmt.Println("[+] All containers have been built!")
 }
 
-// RunDockerComposeStart executes the “docker compose“ commands to start the environment with
-// the specified YAML file (“yaml“ parameter).
+// RunDockerComposeStart executes the "docker compose" commands to start the environment with
+// the specified YAML file ("yaml" parameter).
 func RunDockerComposeStart(yaml string) {
 	fmt.Printf("[+] Running `%s` to restart containers with %s...\n", dockerCmd, yaml)
 	startErr := RunCmd(dockerCmd, []string{"-f", yaml, "start"})
@@ -182,8 +186,8 @@ func RunDockerComposeStart(yaml string) {
 	}
 }
 
-// RunDockerComposeStop executes the “docker compose“ commands to stop all services in the environment with
-// the specified YAML file (“yaml“ parameter).
+// RunDockerComposeStop executes the "docker compose" commands to stop all services in the environment with
+// the specified YAML file ("yaml" parameter).
 func RunDockerComposeStop(yaml string) {
 	fmt.Printf("[+] Running `%s` to stop services with %s...\n", dockerCmd, yaml)
 	stopErr := RunCmd(dockerCmd, []string{"-f", yaml, "stop"})
@@ -192,8 +196,8 @@ func RunDockerComposeStop(yaml string) {
 	}
 }
 
-// RunDockerComposeRestart executes the “docker compose“ commands to restart the environment with
-// the specified YAML file (“yaml“ parameter).
+// RunDockerComposeRestart executes the "docker compose" commands to restart the environment with
+// the specified YAML file ("yaml" parameter).
 func RunDockerComposeRestart(yaml string) {
 	fmt.Printf("[+] Running `%s` to restart containers with %s...\n", dockerCmd, yaml)
 	startErr := RunCmd(dockerCmd, []string{"-f", yaml, "restart"})
@@ -202,8 +206,8 @@ func RunDockerComposeRestart(yaml string) {
 	}
 }
 
-// RunDockerComposeUp executes the “docker compose“ commands to bring up the environment with
-// the specified YAML file (“yaml“ parameter).
+// RunDockerComposeUp executes the "docker compose" commands to bring up the environment with
+// the specified YAML file ("yaml" parameter).
 func RunDockerComposeUp(yaml string) {
 	fmt.Printf("[+] Running `%s` to bring up the containers with %s...\n", dockerCmd, yaml)
 	upErr := RunCmd(dockerCmd, []string{"-f", yaml, "up", "-d"})
@@ -212,8 +216,8 @@ func RunDockerComposeUp(yaml string) {
 	}
 }
 
-// RunDockerComposeDown executes the “docker compose“ commands to bring down the environment with
-// the specified YAML file (“yaml“ parameter).
+// RunDockerComposeDown executes the "docker compose" commands to bring down the environment with
+// the specified YAML file ("yaml" parameter).
 func RunDockerComposeDown(yaml string) {
 	fmt.Printf("[+] Running `%s` to bring down the containers with %s...\n", dockerCmd, yaml)
 	downErr := RunCmd(dockerCmd, []string{"-f", yaml, "down"})
@@ -222,7 +226,7 @@ func RunDockerComposeDown(yaml string) {
 	}
 }
 
-// FetchLogs fetches logs from the container with the specified “name“ label (“containerName“ parameter).
+// FetchLogs fetches logs from the container with the specified "name" label ("containerName" parameter).
 func FetchLogs(containerName string, lines string) []string {
 	var logs []string
 	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
@@ -267,7 +271,7 @@ func FetchLogs(containerName string, lines string) []string {
 	return logs
 }
 
-// GetRunning determines if the container with the specified “name“ label (“containerName“ parameter) is running.
+// GetRunning determines if the container with the specified "name" label ("containerName" parameter) is running.
 func GetRunning() Containers {
 	var running Containers
 
@@ -294,7 +298,7 @@ func GetRunning() Containers {
 	return running
 }
 
-// Determine if the container with the specified “name“ label (“containerName“ parameter) is running.
+// Determine if the container with the specified "name" label ("containerName" parameter) is running.
 func isServiceRunning(containerName string) bool {
 	containers := GetRunning()
 	for _, container := range containers {
@@ -362,7 +366,7 @@ func waitForDjango() bool {
 	}
 }
 
-// RunGhostwriterTests runs Ghostwriter's unit and integration tests via “docker compose“.
+// RunGhostwriterTests runs Ghostwriter's unit and integration tests via "docker compose".
 // The tests are run in the development environment and assume certain values
 // will be set for test conditions, so the .env file is temporarily adjusted
 // during the test run.
@@ -431,19 +435,227 @@ func CheckDockerHealth(dev bool) (HealthIssues, error) {
 	return issues, nil
 }
 
-// RunDockerComposeBackup executes the “docker compose“ command to back up the PostgreSQL database in the environment
-// from the specified YAML file (“yaml“ parameter).
+// generateTime generates a timestamp string for use in backup filenames.
+func generateTime() string {
+	return time.Now().UTC().Format("2006_01_02T15_04_05")
+}
+
+// exists returns whether the given file or directory exists.
+func exists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return false, err
+}
+
+// createTarball creates a tarball from the source directory and writes it to the target file.
+func createTarball(sourceDir string, targetFile string) error {
+	// Open the source directory
+	source, err := os.Open(sourceDir)
+	if err != nil {
+		return err
+	}
+	defer source.Close()
+
+	// Create the target tarball file
+	target, err := os.Create(targetFile)
+	if err != nil {
+		return err
+	}
+	defer target.Close()
+
+	// Create a new gzip writer
+	gzipWriter := gzip.NewWriter(target)
+	defer gzipWriter.Close()
+
+	// Create a new tar writer
+	tarWriter := tar.NewWriter(gzipWriter)
+	defer tarWriter.Close()
+
+	// Walk the source directory
+	err = filepath.Walk(sourceDir, func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+
+		// Create a new tar header
+		header, err := tar.FileInfoHeader(info, info.Name())
+		if err != nil {
+			return err
+		}
+
+		// Adjust the Name field to maintain the file structure in the tarball
+		header.Name, err = filepath.Rel(sourceDir, path)
+		if err != nil {
+			return err
+		}
+
+		// Write the header to the tarball
+		err = tarWriter.WriteHeader(header)
+		if err != nil {
+			return err
+		}
+
+		// If it's not a directory, write the file to the tarball
+		if !info.IsDir() {
+			file, err := os.Open(path)
+			if err != nil {
+				return err
+			}
+			defer file.Close()
+
+			_, err = io.Copy(tarWriter, file)
+			if err != nil {
+				return err
+			}
+		}
+
+		return nil
+	})
+
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// extractTarball extracts the contents of the target tar.gz file.
+func extractTarball(sourceFile string, destDir string) error {
+	// Open the tarball file
+	file, err := os.Open(sourceFile)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+
+	// Create a new gzip reader
+	gzipReader, err := gzip.NewReader(file)
+	if err != nil {
+		return err
+	}
+	defer gzipReader.Close()
+
+	// Create a new tar reader
+	tarReader := tar.NewReader(gzipReader)
+
+	// Iterate over the entries in the tarball
+	for {
+		header, err := tarReader.Next()
+
+		// If no more files are found return
+		if err == io.EOF {
+			break
+		}
+
+		// Return any other error
+		if err != nil {
+			return err
+		}
+
+		// The target location where the dir/file should be created
+		target := filepath.Join(destDir, header.Name)
+
+		// Check the file type
+		switch header.Typeflag {
+
+		// If it's a directory, create it
+		case tar.TypeDir:
+			if _, err := os.Stat(target); err != nil {
+				if err := os.MkdirAll(target, 0755); err != nil {
+					return err
+				}
+			}
+
+		// If it's a file, create it
+		case tar.TypeReg:
+			// Create the file
+			file, err := os.OpenFile(target, os.O_CREATE|os.O_RDWR, os.FileMode(header.Mode))
+			if err != nil {
+				return err
+			}
+			defer file.Close()
+
+			// Copy the file data to the new file
+			if _, err := io.Copy(file, tarReader); err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
+// RunDockerComposeBackup executes the "docker compose" command to back up the PostgreSQL database and related media
+// files in the environment from the specified YAML file ("yaml" parameter).
 func RunDockerComposeBackup(yaml string) {
-	fmt.Printf("[+] Running `%s` to back up the PostgreSQL database with %s...\n", dockerCmd, yaml)
+	// Make a new dump of the PostgreSQL database in teh `production_postgres_data_backups` volume
+	fmt.Printf("[+] Backing up the PostgreSQL database with %s...\n", yaml)
 	backupErr := RunCmd(dockerCmd, []string{"-f", yaml, "run", "--rm", "postgres", "backup"})
 	if backupErr != nil {
 		log.Fatalf("Error trying to back up the PostgreSQL database with %s: %v\n", yaml, backupErr)
 	}
+
+	// Verify there is no "media" directory in the current directory prior to copying the media files
+	preExists, preExistsErr := exists("media")
+	if preExistsErr != nil {
+		log.Fatalf("Error trying to check for the `media` directory: %v\n", preExistsErr)
+	}
+	if preExists {
+		log.Fatalln("A `media` directory already exists in the current directory. Please remove it and try again.")
+	}
+
+	// Copy the Django service's media files locally
+	fmt.Println("[+] Copying media files locally...")
+	cpMediaErr := RunCmd(dockerCmd, []string{"-f", yaml, "cp", "--archive", "django:/app/ghostwriter/media", "."})
+	if cpMediaErr != nil {
+		log.Fatalf("Error trying to copy the media files locally from the Django container with %s: %v\n", yaml, cpMediaErr)
+	}
+
+	// Verify the "media" directory exists after the copy
+	exists, existsErr := exists("media")
+	if existsErr != nil {
+		log.Fatalf("Error trying to check for the `media` directory: %v\n", preExistsErr)
+	}
+	if !exists {
+		log.Fatalln("The copied `media` directory does not exist!")
+	}
+
+	// Create a tarball of the media files
+	fmt.Println("[+] Compressing the media files...")
+	compressFilename := "media_" + generateTime() + ".tar.gz"
+	compErr := createTarball("./media", compressFilename)
+	if compErr != nil {
+		log.Fatalf("Error trying to compress the media files: %v\n", compErr)
+	}
+
+	// Copy the local clone of the Django service's media files to the `production_postgres_data_backups` volume
+	fmt.Println("[+] Copying local media files to the backups volume...")
+	moveMediaErr := RunCmd(dockerCmd, []string{"-f", yaml, "cp", compressFilename, "postgres:/backups"})
+	if moveMediaErr != nil {
+		log.Fatalf("Error trying to move the local media files to the backup volume: %v\n", moveMediaErr)
+	}
+
+	// Delete the local media directory and tarball
+	fmt.Println("[+] Removing the local copies of the media files...")
+	rmDirErr := os.RemoveAll("media")
+	if rmDirErr != nil {
+		log.Fatalf("Error trying to remove the local media files: %v\n", rmDirErr)
+	}
+	rmTarErr := os.Remove(compressFilename)
+	if rmTarErr != nil {
+		log.Fatalf("Error trying to remove the local media files: %v\n", rmTarErr)
+	}
+	fmt.Println("[+] Done!")
 }
 
-// RunDockerComposeBackups executes the “docker compose“ command to list available PostgreSQL database backups in the
-// environment from the specified YAML file (“yaml“ parameter).
-func RunDockerComposeBackups(yaml string) {
+// RunDockerListBackups executes the "docker compose" command to list available backup archives in the
+// environment from the specified YAML file ("yaml" parameter).
+func RunDockerListBackups(yaml string) {
 	fmt.Printf("[+] Running `%s` to list avilable PostgreSQL database backup files with %s...\n", dockerCmd, yaml)
 	backupErr := RunCmd(dockerCmd, []string{"-f", yaml, "run", "--rm", "postgres", "backups"})
 	if backupErr != nil {
@@ -451,12 +663,84 @@ func RunDockerComposeBackups(yaml string) {
 	}
 }
 
-// RunDockerComposeRestore executes the “docker compose“ command to restore a PostgreSQL database backup in the
-// environment from the specified YAML file (“yaml“ parameter).
-func RunDockerComposeRestore(yaml string, restore string) {
-	fmt.Printf("[+] Running `%s` to restore the PostgreSQL database backup file %s with %s...\n", dockerCmd, restore, yaml)
-	backupErr := RunCmd(dockerCmd, []string{"-f", yaml, "run", "--rm", "postgres", "restore", restore})
-	if backupErr != nil {
-		log.Fatalf("Error trying to restore %s with %s: %v\n", restore, yaml, backupErr)
+// RunDockerDownloadBackups executes the "docker cp" command to copy backup files to the local `backups` directory.
+func RunDockerDownloadBackups(yaml string) {
+	// Copy the contents of the `production_postgres_data_backups` volume to the current directory
+	fmt.Println("[+] Copying backup files locally...")
+	cpSqlErr := RunCmd(dockerCmd, []string{"-f", yaml, "cp", "postgres:/backups", "."})
+	if cpSqlErr != nil {
+		log.Fatalf("Error trying to copy the backup files from the PostgreSQL container with %s: %v\n", yaml, cpSqlErr)
 	}
+
+	// Verify the "backups" directory exists after the copy
+	exists, existsErr := exists("backups")
+	if existsErr != nil {
+		log.Fatalf("Error trying to check for the `backups` directory: %v\n", existsErr)
+	}
+	if !exists {
+		log.Fatalln("The copied `backups` directory does not exist")
+	}
+}
+
+// RunDockerComposeRestore executes the "docker compose" command to restore a PostgreSQL database backup in the
+// environment from the specified YAML file ("yaml" parameter).
+func RunDockerComposeRestore(yaml string, dbRestore string, mediaRestore string) {
+	// Verify a `media` directory does not already exist
+	fmt.Println("[+] Checking for a pre-existing `media` directory that might be overwritten")
+	dirExists, dirExistsErr := exists("media")
+	if dirExistsErr != nil {
+		log.Fatalf("Error trying to check for an existing `media` directory: %v\n", dirExistsErr)
+	}
+	if dirExists {
+		log.Fatalln("A `media` directory already exists in the current directory. Please remove it and try again.")
+	}
+
+	// Restore the SQL dump
+	fmt.Printf("[+] Running `%s` to restore the PostgreSQL database backup file %s with %s...\n", dockerCmd, dbRestore, yaml)
+	backupErr := RunCmd(dockerCmd, []string{"-f", yaml, "run", "--rm", "postgres", "restore", dbRestore})
+	if backupErr != nil {
+		log.Fatalf("Error trying to restore %s with %s: %v\n", dbRestore, yaml, backupErr)
+	}
+
+	// Download the media files from the backup volume
+	fmt.Printf("[+] Copying media backup, %s, file locally...\n", mediaRestore)
+	cpErr := RunCmd(dockerCmd, []string{"-f", yaml, "cp", "postgres:/backups/" + mediaRestore, "."})
+	if cpErr != nil {
+		log.Fatalf("Error trying to copy the %s with %s: %v\n", mediaRestore, yaml, cpErr)
+	}
+
+	// Verify the "media" directory exists after the copy
+	tarExists, existsErr := exists(mediaRestore)
+	if existsErr != nil {
+		log.Fatalf("Error trying to check for the %s archive: %v\n", mediaRestore, existsErr)
+	}
+	if !tarExists {
+		log.Fatalf("The copied %s archive does not exist!\n", mediaRestore)
+	}
+
+	// Extract the contents of the media archive
+	err := extractTarball(mediaRestore, "./media")
+	if err != nil {
+		log.Fatalf("Error trying to extract the archive: %v\n", err)
+	}
+
+	// Copy the media files to the Django service's media directory
+	fmt.Println("[+] Copying media files to the container's volume...")
+	cpMediaErr := RunCmd(dockerCmd, []string{"-f", yaml, "cp", "./media", "django:/app/ghostwriter/"})
+	if cpMediaErr != nil {
+		log.Fatalf("Error trying to copy the contents of %s to the volume with %s: %v\n", mediaRestore, yaml, cpMediaErr)
+	}
+
+	// Delete the local media directory
+	fmt.Println("[+] Removing the local copies of the media files...")
+	rmDirErr := os.RemoveAll("media")
+	if rmDirErr != nil {
+		log.Fatalf("Error trying to remove the local media files: %v\n", rmDirErr)
+	}
+	rmTarErr := os.Remove(mediaRestore)
+	if rmTarErr != nil {
+		log.Fatalf("Error trying to remove the local copy of %s: %v\n", mediaRestore, rmTarErr)
+	}
+
+	fmt.Println("[+] Done!")
 }